refactor(ricer::tests)!: combine unit tests to match current module structure
