+++ src/tests/cli.rs
