+++ src/tests/config.rs
@@ -1,16 +1,246 @@
 // SPDX-FileCopyrightText: 2024 Jason Pena <jasonpena@awkless.com>
 // SPDX-License-Identifier: MIT
 
-use crate::config::{
-    CmdHookSettings, CmdHookConfig, ConfigFile, ConfigFileError, HookSettings, RepoSettings,
-    RepoConfig, Settings, Config,
-};
+use crate::config::*;
 use crate::locate::MockLocator;
 use crate::tests::FakeConfigDir;
 
 use anyhow::Result;
-use indoc::indoc; use pretty_assertions::assert_eq;
-use rstest::rstest;
+use indoc::{formatdoc, indoc};
+use pretty_assertions::assert_eq;
+use rstest::{fixture, rstest};
+use toml_edit::{DocumentMut, Item, Key, Value};
+
+fn setup_repo_doc(entry: (Key, Item)) -> Result<DocumentMut> {
+    let mut doc: DocumentMut = "[repos]".parse()?;
+    let table = doc.get_mut("repos").unwrap();
+    let table = table.as_table_mut().unwrap();
+    let (key, item) = entry;
+    table.insert_formatted(&key, item);
+    table.set_implicit(true);
+    Ok(doc)
+}
+
+fn setup_cmd_hook_doc(entry: (Key, Item)) -> Result<DocumentMut> {
+    let mut doc: DocumentMut = "[hooks]".parse()?;
+    let table = doc.get_mut("hooks").unwrap();
+    let table = table.as_table_mut().unwrap();
+    let (key, item) = entry;
+    table.insert_formatted(&key, item);
+    table.set_implicit(true);
+    Ok(doc)
+}
+
+#[fixture]
+fn toml_input() -> String {
+    String::from(indoc! {r#"
+        # this coment should remain!
+        [test]
+        foo = "hello"
+        bar = true
+    "#})
+}
+
+#[rstest]
+fn toml_parse_accepts_good_formatting(
+    #[values("this = 'will parse'", "[so_will_this]", "hello.world = 'from ricer!'")] input: &str,
+) -> Result<()> {
+    let toml: Result<Toml, TomlError> = input.parse();
+    assert!(toml.is_ok());
+    Ok(())
+}
+
+#[rstest]
+fn toml_parse_catches_bad_formatting(
+    #[values("this 'will fail'", "[will # also fail", "not.gonna = [work]")] input: &str,
+) {
+    let result: Result<Toml, TomlError> = input.parse();
+    assert!(matches!(result.unwrap_err(), TomlError::BadParse { .. }));
+}
+
+#[rstest]
+#[case("test", "foo", (Key::new("foo"), Item::Value(Value::from("hello"))))]
+#[case("test", "bar", (Key::new("bar"), Item::Value(Value::from(true))))]
+fn toml_get_returns_entry(
+    toml_input: String,
+    #[case] table: &str,
+    #[case] key: &str,
+    #[case] expect: (Key, Item),
+) -> Result<()> {
+    let toml: Toml = toml_input.parse()?;
+    let (result_key, result_value) = toml.get(table, key)?;
+    let (expect_key, expect_value) = expect;
+    assert_eq!(result_key, &expect_key);
+    assert_eq!(result_value.is_value(), expect_value.is_value());
+    Ok(())
+}
+
+#[rstest]
+#[case::table_not_found("bar = 'foo not here'", TomlError::TableNotFound { table: "foo".into() })]
+#[case::not_table("foo = 'not a table'", TomlError::NotTable { table: "foo".into() })]
+#[case::entry_not_found(
+    "[foo] # bar not here",
+    TomlError::EntryNotFound { table: "foo".into(), key: "bar".into() }
+)]
+fn toml_get_catches_errors(#[case] input: &str, #[case] expect: TomlError) -> Result<()> {
+    let toml: Toml = input.parse()?;
+    let result = toml.get("foo", "bar");
+    assert_eq!(result.unwrap_err(), expect);
+    Ok(())
+}
+
+#[rstest]
+#[case::add_into_table(
+    toml_input(),
+    "test",
+    (Key::new("baz"), Item::Value(Value::from("add this"))),
+    formatdoc! {r#"
+        {}baz = "add this"
+    "#, toml_input()}
+)]
+#[case::create_new_table(
+    toml_input(),
+    "new_test",
+    (Key::new("baz"), Item::Value(Value::from("add this"))),
+    formatdoc! {r#"
+        {}
+        [new_test]
+        baz = "add this"
+    "#, toml_input()}
+)]
+fn toml_add_new_entry(
+    #[case] input: String,
+    #[case] table: &str,
+    #[case] entry: (Key, Item),
+    #[case] expect: String,
+) -> Result<()> {
+    let mut toml: Toml = input.parse()?;
+    let result = toml.add(table, entry)?;
+    assert_eq!(toml.to_string(), expect);
+    assert!(result.is_none());
+    Ok(())
+}
+
+#[rstest]
+#[case(
+    toml_input(),
+    "test",
+    (Key::new("foo"), Item::Value(Value::from("replaced"))),
+    toml_input().replace(r#"foo = "hello""#, r#"foo = "replaced""#)
+)]
+#[case(
+    toml_input(),
+    "test",
+    (Key::new("bar"), Item::Value(Value::from(false))),
+    toml_input().replace(r#"bar = true"#, r#"bar = false"#)
+)]
+fn toml_add_replace_entry(
+    #[case] input: String,
+    #[case] table: &str,
+    #[case] entry: (Key, Item),
+    #[case] expect: String,
+) -> Result<()> {
+    let mut toml: Toml = input.parse()?;
+    let result = toml.add(table, entry)?;
+    assert_eq!(toml.to_string(), expect);
+    assert!(result.is_some());
+    Ok(())
+}
+
+#[rstest]
+#[case::not_table("foo = 'not a table'", TomlError::NotTable { table: "foo".into() })]
+fn toml_add_catches_errors(#[case] input: &str, #[case] expect: TomlError) -> Result<()> {
+    let mut toml: Toml = input.parse()?;
+    let stub = (Key::new("fail"), Item::Value(Value::from("this")));
+    let result = toml.add("foo", stub);
+    assert_eq!(result.unwrap_err(), expect);
+    Ok(())
+}
+
+#[rstest]
+#[case(
+    toml_input(),
+    "test",
+    "bar",
+    "baz",
+    (Key::new("bar"), Item::Value(Value::from(true))),
+    toml_input().replace("bar", "baz"),
+)]
+fn toml_rename_renames_entry(
+    #[case] input: String,
+    #[case] table: &str,
+    #[case] from: &str,
+    #[case] to: &str,
+    #[case] expect: (Key, Item),
+    #[case] output: String,
+) -> Result<()> {
+    let mut toml: Toml = input.parse()?;
+    let (return_key, return_value) = toml.rename(table, from, to)?;
+    let (expect_key, expect_value) = expect;
+    assert_eq!(toml.to_string(), output);
+    assert_eq!(return_key, expect_key);
+    assert_eq!(return_value.is_value(), expect_value.is_value());
+    Ok(())
+}
+
+#[rstest]
+#[case::table_not_found("bar = 'foo not here'", TomlError::TableNotFound { table: "foo".into() })]
+#[case::not_table("foo = 'not a table'", TomlError::NotTable { table: "foo".into() })]
+#[case::entry_not_found(
+    "[foo] # bar not here",
+    TomlError::EntryNotFound { table: "foo".into(), key: "bar".into() }
+)]
+fn toml_rename_catches_errors(#[case] input: &str, #[case] expect: TomlError) -> Result<()> {
+    let toml: Toml = input.parse()?;
+    let result = toml.get("foo", "bar");
+    assert_eq!(result.unwrap_err(), expect);
+    Ok(())
+}
+
+#[rstest]
+#[case(
+    toml_input(),
+    "test",
+    "foo",
+    (Key::new("foo"), Item::Value(Value::from("world"))),
+    toml_input().replace("foo = \"hello\"\n", ""),
+)]
+#[case(
+    toml_input(),
+    "test",
+    "bar",
+    (Key::new("bar"), Item::Value(Value::from(true))),
+    toml_input().replace("bar = true\n", ""),
+)]
+fn toml_remove_deletes_entry(
+    #[case] input: String,
+    #[case] table: &str,
+    #[case] key: &str,
+    #[case] expect: (Key, Item),
+    #[case] output: String,
+) -> Result<()> {
+    let mut toml: Toml = input.parse()?;
+    let (return_key, return_value) = toml.remove(table, key)?;
+    let (expect_key, expect_value) = expect;
+    assert_eq!(toml.to_string(), output);
+    assert_eq!(return_key, expect_key);
+    assert_eq!(return_value.is_value(), expect_value.is_value());
+    Ok(())
+}
+
+#[rstest]
+#[case::table_not_found("bar = 'foo not here'", TomlError::TableNotFound { table: "foo".into() })]
+#[case::not_table("foo = 'not a table'", TomlError::NotTable { table: "foo".into() })]
+#[case::entry_not_found(
+    "[foo] # bar not here",
+    TomlError::EntryNotFound { table: "foo".into(), key: "bar".into() }
+)]
+fn toml_remove_catches_errors(#[case] input: &str, #[case] expect: TomlError) -> Result<()> {
+    let toml: Toml = input.parse()?;
+    let result = toml.get("foo", "bar");
+    assert_eq!(result.unwrap_err(), expect);
+    Ok(())
+}
 
 #[rstest]
 #[case::repo_data(
@@ -21,7 +251,7 @@ use rstest::rstest;
     CmdHookConfig,
     FakeConfigDir::builder()?.config_file("hooks.toml", "this = 'will parse'\n")?.build(),
 )]
-fn config_manager_load_works(
+fn config_file_load_works(
     #[case] config_type: impl Config,
     #[case] config_data: FakeConfigDir,
 ) -> Result<()> {
@@ -44,7 +274,7 @@ fn config_manager_load_works(
     CmdHookConfig,
     FakeConfigDir::builder()?.config_file("hooks.toml", "this 'will fail'")?.build(),
 )]
-fn config_manager_load_catches_toml_error(
+fn config_file_load_catches_toml_error(
     #[case] config_type: impl Config,
     #[case] config_data: FakeConfigDir,
 ) -> Result<()> {
@@ -61,7 +291,7 @@ fn config_manager_load_catches_toml_error(
 #[rstest]
 #[case::repo_data(RepoConfig, FakeConfigDir::builder()?.build())]
 #[case::hook_data(CmdHookConfig, FakeConfigDir::builder()?.build())]
-fn config_manager_load_creates_new_file(
+fn config_file_load_creates_new_file(
     #[case] config_type: impl Config,
     #[case] config_data: FakeConfigDir,
 ) -> Result<()> {
@@ -106,7 +336,7 @@ fn config_manager_load_creates_new_file(
     )?.build(),
     CmdHookSettings::new("commit").add_hook(HookSettings::new().post("hook.sh")),
 )]
-fn config_manager_save_works<E, T>(
+fn config_file_save_works<E, T>(
     #[case] config_type: T,
     #[case] mut config_data: FakeConfigDir,
     #[case] entry: E,
@@ -131,7 +361,7 @@ where
 #[rstest]
 #[case::repo_data(RepoConfig, FakeConfigDir::builder()?.build())]
 #[case::hook_data(CmdHookConfig, FakeConfigDir::builder()?.build())]
-fn config_manager_save_creates_new_file(
+fn config_file_save_creates_new_file(
     #[case] config_type: impl Config,
     #[case] config_data: FakeConfigDir,
 ) -> Result<()> {
@@ -178,7 +408,7 @@ fn config_manager_save_creates_new_file(
         .add_hook(HookSettings::new().pre("hook.sh").post("hook.sh").workdir("/some/dir"))
         .add_hook(HookSettings::new().pre("hook.sh")),
 )]
-fn config_manager_get_works<E, T>(
+fn config_file_get_works<E, T>(
     #[case] config_type: T,
     #[case] config_data: FakeConfigDir,
     #[case] key: &str,
@@ -202,7 +432,7 @@ where
 #[rstest]
 #[case(RepoConfig, FakeConfigDir::builder()?.build())]
 #[case(CmdHookConfig, FakeConfigDir::builder()?.build())]
-fn config_manager_get_catches_errors(
+fn config_file_get_catches_errors(
     #[case] config_type: impl Config,
     #[case] config_data: FakeConfigDir,
 ) -> Result<()> {
@@ -243,7 +473,7 @@ fn config_manager_get_catches_errors(
         ]
     "#},
 )]
-fn config_manager_new_data<E, T>(
+fn config_file_new_data<E, T>(
     #[case] config_type: T,
     #[case] config_data: FakeConfigDir,
     #[case] entry: E,
@@ -279,7 +509,7 @@ where
         .build(),
     CmdHookSettings::default(),
 )]
-fn config_manager_add_catches_errors<E, T>(
+fn config_file_add_catches_errors<E, T>(
     #[case] config_type: T,
     #[case] config_data: FakeConfigDir,
     #[case] entry: E,
@@ -350,7 +580,7 @@ where
         ]
     "#}
 )]
-fn config_manager_rename_works<E, T>(
+fn config_file_rename_works<E, T>(
     #[case] config_type: T,
     #[case] config_data: FakeConfigDir,
     #[case] from: &str,
@@ -377,7 +607,7 @@ where
 #[rstest]
 #[case::repo_data(RepoConfig, FakeConfigDir::builder()?.build())]
 #[case::hook_data(CmdHookConfig, FakeConfigDir::builder()?.build())]
-fn config_manager_rename_catches_errors(
+fn config_file_rename_catches_errors(
     #[case] config_type: impl Config,
     #[case] config_data: FakeConfigDir,
 ) -> Result<()> {
@@ -449,7 +679,7 @@ fn config_manager_rename_catches_errors(
             ]
     "#},
 )]
-fn config_manager_remove_works<E, T>(
+fn config_file_remove_works<E, T>(
     #[case] config_type: T,
     #[case] config_data: FakeConfigDir,
     #[case] key: &str,
@@ -475,7 +705,7 @@ where
 #[rstest]
 #[case::repo_data(RepoConfig, FakeConfigDir::builder()?.build())]
 #[case::hook_data(CmdHookConfig, FakeConfigDir::builder()?.build())]
-fn config_manager_remove_catches_errors(
+fn config_file_remove_catches_errors(
     #[case] config_type: impl Config,
     #[case] config_data: FakeConfigDir,
 ) -> Result<()> {
@@ -487,5 +717,115 @@ fn config_manager_remove_catches_errors(
     let result = config.remove("non-existent");
     assert!(matches!(result.unwrap_err(), ConfigFileError::Toml { .. }));
 
+    Ok(())
+}
+
+#[rstest]
+#[case::no_bootstrap(
+    RepoSettings::new("vim")
+        .branch("master")
+        .remote("origin")
+        .workdir_home(true),
+    indoc! {r#"
+        [repos.vim]
+        branch = "master"
+        remote = "origin"
+        workdir_home = true
+    "#}
+)]
+#[case::with_bootstrap(
+    RepoSettings::new("vim")
+        .branch("master")
+        .remote("origin")
+        .workdir_home(true)
+        .bootstrap(
+            BootstrapSettings::new()
+                .clone("https://github.com/awkless/vim.git")
+                .os(OsType::Unix)
+                .users(["awkless", "sedgwick"])
+                .hosts(["lovelace", "turing"])
+        ),
+    indoc! {r#"
+        [repos.vim]
+        branch = "master"
+        remote = "origin"
+        workdir_home = true
+
+        [repos.vim.bootstrap]
+        clone = "https://github.com/awkless/vim.git"
+        os = "unix"
+        users = ["awkless", "sedgwick"]
+        hosts = ["lovelace", "turing"]
+    "#}
+)]
+fn repo_settings_to_toml_serialize(#[case] repo: RepoSettings, #[case] expect: &str) -> Result<()> {
+    let doc = setup_repo_doc(repo.to_toml())?;
+    assert_eq!(doc.to_string(), expect);
+    Ok(())
+}
+
+#[rstest]
+#[case::no_bootstrap(
+    RepoSettings::new("vim")
+        .branch("master")
+        .remote("origin")
+        .workdir_home(true),
+)]
+#[case::with_bootstrap(
+    RepoSettings::new("vim")
+        .branch("master")
+        .remote("origin")
+        .workdir_home(true)
+        .bootstrap(
+            BootstrapSettings::new()
+                .clone("https://github.com/awkless/vim.git")
+                .os(OsType::Unix)
+                .users(["awkless", "sedgwick"])
+                .hosts(["lovelace", "turing"])
+        ),
+)]
+fn repo_settings_from_entry_deserialize(#[case] expect: RepoSettings) -> Result<()> {
+    let doc = setup_repo_doc(expect.to_toml())?;
+    let result = RepoSettings::from(doc["repos"].as_table().unwrap().get_key_value("vim").unwrap());
+    assert_eq!(result, expect);
+    Ok(())
+}
+
+#[rstest]
+#[case(
+    CmdHookSettings::new("commit")
+        .add_hook(HookSettings::new().pre("hook.sh").post("hook.sh").workdir("/some/path"))
+        .add_hook(HookSettings::new().pre("hook.sh"))
+        .add_hook(HookSettings::new().post("hook.sh")),
+    indoc! {r#"
+        [hooks]
+        commit = [
+            { pre = "hook.sh", post = "hook.sh", workdir = "/some/path" },
+            { pre = "hook.sh" },
+            { post = "hook.sh" }
+        ]
+    "#}
+)]
+fn cmd_hook_settings_to_toml_serialize(
+    #[case] cmd_hook: CmdHookSettings,
+    #[case] expect: &str,
+) -> Result<()> {
+    let doc = setup_cmd_hook_doc(cmd_hook.to_toml())?;
+    assert_eq!(doc.to_string(), expect);
+    Ok(())
+}
+
+#[rstest]
+#[case(
+    CmdHookSettings::new("commit")
+        .add_hook(HookSettings::new().pre("hook.sh").post("hook.sh").workdir("/some/path"))
+        .add_hook(HookSettings::new().pre("hook.sh"))
+        .add_hook(HookSettings::new().post("hook.sh"))
+)]
+fn cmd_hook_settings_from_deserialize(#[case] expect: CmdHookSettings) -> Result<()> {
+    let doc = setup_cmd_hook_doc(expect.to_toml())?;
+    let result =
+        CmdHookSettings::from(doc["hooks"].as_table().unwrap().get_key_value("commit").unwrap());
+    assert_eq!(result, expect);
     Ok(())
 }
